[M1] refactor: 답글 삭제 가능 여부 판단 역할을 답글 클래스로 이동
참고: https://github.com/next-step/java-bowling/pull/885
